+++ Tests/SentryTests/SentryInterfacesTests.m
@@ -1,5 +1,6 @@
 #import <XCTest/XCTest.h>
 
+#import "SentryBreadcrumb+Private.h"
 #import "SentryBreadcrumb.h"
 #import "SentryDateUtils.h"
 #import "SentryEvent.h"
@@ -378,10 +379,12 @@
     crumb2.type = @"type";
     crumb2.timestamp = date;
     crumb2.message = @"message";
+    crumb2.origin = @"origin";
     NSDictionary *serialized2 = @{
         @"level" : @"info",
         @"type" : @"type",
         @"message" : @"message",
+        @"origin" : @"origin",
         @"timestamp" : sentry_toIso8601String(date),
         @"category" : @"http",
         @"data" : @ { @"bla" : @"1" },